% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life2010.R
\docType{data}
\name{life2010}
\alias{life2010}
\title{Lifetime table of Korea 2010}
\format{
A data frame with 101 rows and 3 variables:
\describe{
    \item{\code{Age}}{age}
    \item{\code{Prob_d_m}}{probability of death for male}
    \item{\code{Prob_d_f}}{probability of death for female}
}
}
\source{
KOSIS(Statistics Korea, Life Tables By Province) \url{https://kosis.kr/}
}
\usage{
life2010
}
\description{
A dataset containing the probability of death by age and gender.
}
\keyword{datasets}
