% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuclear.R
\docType{data}
\name{nuclear}
\alias{nuclear}
\title{Simulated data of organ radiation exposure dose}
\format{
A data frame with 100 observation of 11 variables:
\describe{
    \item{\code{ID}}{person ID.}
    \item{\code{sex}}{gender}
    \item{\code{birth}}{birth-year}
    \item{\code{exposure}}{exposed year to radiation}
    \item{\code{site}}{organ where exposed to radiation}
    \item{\code{exposure_rate}}{expsoure rate}
    \item{\code{dosedist}}{distribution of dose}
    \item{\code{dose1}}{dose parameter}
    \item{\code{dose2}}{dose parameter}
    \item{\code{dose3}}{dose parameter}
    \item{\code{distance}}{distance from the hyper}
}
}
\usage{
nuclear
}
\description{
\code{nuclear} is simulated dataset for acute exposure event. The scenario assumes that the people exposure the radiation at 2011.
}
\keyword{datasets}
