% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organ.R
\docType{data}
\name{organ}
\alias{organ}
\title{Simulated data of organ radiation exposure dose}
\format{
A data frame with 971 observation of 11 variables:
\describe{
    \item{\code{ID}}{person ID.}
    \item{\code{sex}}{gender}
    \item{\code{birth}}{birth-year}
    \item{\code{exposure}}{exposed year to radiation}
    \item{\code{site}}{organ where exposed to radiation}
    \item{\code{exposure_rate}}{expsoure rate}
    \item{\code{dosedist}}{distribution of dose}
    \item{\code{dose1}}{dose parameter}
    \item{\code{dose2}}{dose parameter}
    \item{\code{dose3}}{dose parameter}
    \item{\code{occup}}{occupation}
}
}
\usage{
organ
}
\description{
\code{organ} is simulated dataset from the data of workers at interventional radiology departments.
}
\references{
Lee, W. J., Bang, Y. J., Cha, E. S., Kim, Y. M., & Cho, S. B.
 (2021). Lifetime cancer risks from occupational radiation exposure among
 workers at interventional radiology departments. \emph{International
 Archives of Occupational and Environmental Health}, \bold{94(1)}, 139-145.
}
\keyword{datasets}
