% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LAR.R
\name{print.LAR}
\alias{print.LAR}
\alias{print.LAR_batch}
\alias{print.LAR_group}
\title{Print estimated Lifetime Attributable Risk for one person}
\usage{
\method{print}{LAR}(x, digits = 4, ...)

\method{print}{LAR_batch}(x, digits = 4, max.id = 50, ...)

\method{print}{LAR_group}(x, digits = 4, max.id = 50, ...)
}
\arguments{
\item{x}{'LAR', 'LAR_batch' or 'LAR_group' object.}

\item{digits}{the number of decimal points to print.}

\item{...}{further arguments to be passed from or to other methods.}

\item{max.id}{the number of maximum of printing LAR results.}
}
\description{
\code{print.LAR} is the basic function for printing class "LAR".
}
