% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LAR.R
\name{summary.LAR}
\alias{summary.LAR}
\alias{summary.LAR_batch}
\alias{summary.LAR_group}
\title{Summarize estimated Lifetime Attributable Risk for one person}
\usage{
\method{summary}{LAR}(object, digits = 4, ...)

\method{summary}{LAR_batch}(object, digits = 4, max.id = 50, ...)

\method{summary}{LAR_group}(object, digits = 4, max.id = 50, ...)
}
\arguments{
\item{object}{object of class 'LAR_batch' or LAR'.}

\item{digits}{the number of decimal points to print.}

\item{...}{further arguments to be passed from or to other methods.}

\item{max.id}{the number of maximum of printing LAR results.}
}
\description{
\code{summary.LAR} is the function for printing class "LAR".
}
