% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_LAR.R
\name{write_LAR}
\alias{write_LAR}
\alias{write_LAR.LAR}
\alias{write_LAR.LAR_batch}
\alias{write_LAR.LAR_group}
\title{Write a LAR object}
\usage{
write_LAR(x, filename)

\method{write_LAR}{LAR}(x, filename)

\method{write_LAR}{LAR_batch}(x, filename)

\method{write_LAR}{LAR_group}(x, filename)
}
\arguments{
\item{x}{a 'LAR' object.}

\item{filename}{a string naming the file to save (.csv file)}
}
\description{
Write 'LAR' object to CSV file
}
\section{Methods (by class)}{
\itemize{
\item \code{LAR}: write an 'LAR' class object

\item \code{LAR_batch}: write an 'LAR_batch' class object

\item \code{LAR_group}: write an 'LAR_group' class object
}}

